Introduce sources package
