+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/analyzer/CustomEventAnalyzer.scala
@@ -3,7 +3,7 @@ package com.microsoft.partnercatalyst.fortis.spark.analyzer
 import java.time.Instant.now
 import java.util.UUID.randomUUID
 
-import com.microsoft.partnercatalyst.fortis.spark.streamwrappers.customevents.CustomEvent
+import com.microsoft.partnercatalyst.fortis.spark.sources.streamwrappers.customevents.CustomEvent
 import com.microsoft.partnercatalyst.fortis.spark.transforms.image.ImageAnalyzer
 
 @SerialVersionUID(100L)
