+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/analyzer/RadioAnalyzer.scala
@@ -3,7 +3,7 @@ package com.microsoft.partnercatalyst.fortis.spark.analyzer
 import java.time.Instant.now
 import java.util.UUID.randomUUID
 
-import com.microsoft.partnercatalyst.fortis.spark.streamwrappers.radio.RadioTranscription
+import com.microsoft.partnercatalyst.fortis.spark.sources.streamwrappers.radio.RadioTranscription
 import com.microsoft.partnercatalyst.fortis.spark.transforms.image.ImageAnalyzer
 
 @SerialVersionUID(100L)
