+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/analyzer/TadawebAnalyzer.scala
@@ -3,7 +3,7 @@ package com.microsoft.partnercatalyst.fortis.spark.analyzer
 import java.time.Instant.now
 import java.util.UUID.randomUUID
 
-import com.microsoft.partnercatalyst.fortis.spark.streamwrappers.tadaweb.TadawebEvent
+import com.microsoft.partnercatalyst.fortis.spark.sources.streamwrappers.tadaweb.TadawebEvent
 import com.microsoft.partnercatalyst.fortis.spark.transforms.image.ImageAnalyzer
 import com.microsoft.partnercatalyst.fortis.spark.transforms.sentiment.SentimentDetector
