+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/dba/ConfigurationManager.scala
@@ -1,7 +1,7 @@
 package com.microsoft.partnercatalyst.fortis.spark.dba
 
 import com.microsoft.partnercatalyst.fortis.spark.dto.SiteSettings
-import com.microsoft.partnercatalyst.fortis.spark.streamprovider.ConnectorConfig
+import com.microsoft.partnercatalyst.fortis.spark.sources.streamprovider.ConnectorConfig
 
 trait ConfigurationManager {
   def fetchConnectorConfigs(pipeline: String): List[ConnectorConfig]
