+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/serialization/KryoRegistrator.scala
@@ -41,10 +41,10 @@ class KryoRegistrator extends BaseKryoRegistrator {
     kryo.register(classOf[facebook4j.Comment])
     kryo.register(classOf[facebook4j.Application])
     kryo.register(classOf[facebook4j.Targeting])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.tadaweb.TadawebEvent])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.tadaweb.TadawebCity])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.tadaweb.TadawebTada])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.radio.RadioTranscription])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.sources.streamwrappers.tadaweb.TadawebEvent])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.sources.streamwrappers.tadaweb.TadawebCity])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.sources.streamwrappers.tadaweb.TadawebTada])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.sources.streamwrappers.radio.RadioTranscription])
     kryo.register(classOf[com.github.catalystcode.fortis.spark.streaming.instagram.dto.InstagramItem])
     kryo.register(classOf[com.github.catalystcode.fortis.spark.streaming.instagram.dto.InstagramUser])
     kryo.register(classOf[com.github.catalystcode.fortis.spark.streaming.instagram.dto.InstagramLocation])
@@ -63,9 +63,9 @@ class KryoRegistrator extends BaseKryoRegistrator {
     kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.analyzer.TwitterAnalyzer])
     kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.locations.LocationsExtractorFactory])
     kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.transforms.ZipModelsProvider])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.customevents.CustomEvent])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.customevents.CustomEventFeature])
-    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.streamwrappers.customevents.CustomEventFeatureCollection])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.sources.streamwrappers.customevents.CustomEvent])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.sources.streamwrappers.customevents.CustomEventFeature])
+    kryo.register(classOf[com.microsoft.partnercatalyst.fortis.spark.sources.streamwrappers.customevents.CustomEventFeatureCollection])
 
     kryo.register(classOf[com.github.catalystcode.fortis.spark.streaming.reddit.dto.RedditObject])
     kryo.register(classOf[com.github.catalystcode.fortis.spark.streaming.reddit.dto.RedditObjectData])
