+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/sources/streamfactories/InstagramLocationStreamFactory.scala
@@ -1,12 +1,12 @@
-package com.microsoft.partnercatalyst.fortis.spark.streamfactories
+package com.microsoft.partnercatalyst.fortis.spark.sources.streamfactories
 
 import com.github.catalystcode.fortis.spark.streaming.instagram.dto.InstagramItem
 import com.github.catalystcode.fortis.spark.streaming.instagram.{InstagramAuth, InstagramUtils}
-import com.microsoft.partnercatalyst.fortis.spark.streamprovider.{ConnectorConfig, StreamFactory}
+import com.microsoft.partnercatalyst.fortis.spark.sources.streamprovider.{ConnectorConfig, StreamFactory}
 import org.apache.spark.streaming.StreamingContext
 import org.apache.spark.streaming.dstream.DStream
 
-class InstagramTagStreamFactory extends StreamFactory[InstagramItem]{
+class InstagramLocationStreamFactory extends StreamFactory[InstagramItem]{
   /**
     * Creates a DStream for a given connector config iff the connector config is supported by the stream factory.
     * The param set allows the streaming context to be curried into the partial function which creates the stream.
@@ -15,9 +15,13 @@ class InstagramTagStreamFactory extends StreamFactory[InstagramItem]{
     * @return A partial function for transforming a connector config
     */
   override def createStream(streamingContext: StreamingContext): PartialFunction[ConnectorConfig, DStream[InstagramItem]] = {
-    case ConnectorConfig("InstagramTag", params) =>
+    case ConnectorConfig("InstagramLocation", params) =>
       val auth = InstagramAuth(params("authToken"))
 
-      InstagramUtils.createTagStream(streamingContext, auth, params("tag"))
+      InstagramUtils.createLocationStream(
+        streamingContext,
+        auth,
+        latitude = params("latitude").toDouble,
+        longitude = params("longitude").toDouble)
   }
 }
