+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/sources/streamfactories/InstagramTagStreamFactory.scala
@@ -1,12 +1,12 @@
-package com.microsoft.partnercatalyst.fortis.spark.streamfactories
+package com.microsoft.partnercatalyst.fortis.spark.sources.streamfactories
 
 import com.github.catalystcode.fortis.spark.streaming.instagram.dto.InstagramItem
 import com.github.catalystcode.fortis.spark.streaming.instagram.{InstagramAuth, InstagramUtils}
-import com.microsoft.partnercatalyst.fortis.spark.streamprovider.{ConnectorConfig, StreamFactory}
+import com.microsoft.partnercatalyst.fortis.spark.sources.streamprovider.{ConnectorConfig, StreamFactory}
 import org.apache.spark.streaming.StreamingContext
 import org.apache.spark.streaming.dstream.DStream
 
-class InstagramLocationStreamFactory extends StreamFactory[InstagramItem]{
+class InstagramTagStreamFactory extends StreamFactory[InstagramItem]{
   /**
     * Creates a DStream for a given connector config iff the connector config is supported by the stream factory.
     * The param set allows the streaming context to be curried into the partial function which creates the stream.
@@ -15,13 +15,9 @@ class InstagramLocationStreamFactory extends StreamFactory[InstagramItem]{
     * @return A partial function for transforming a connector config
     */
   override def createStream(streamingContext: StreamingContext): PartialFunction[ConnectorConfig, DStream[InstagramItem]] = {
-    case ConnectorConfig("InstagramLocation", params) =>
+    case ConnectorConfig("InstagramTag", params) =>
       val auth = InstagramAuth(params("authToken"))
 
-      InstagramUtils.createLocationStream(
-        streamingContext,
-        auth,
-        latitude = params("latitude").toDouble,
-        longitude = params("longitude").toDouble)
+      InstagramUtils.createTagStream(streamingContext, auth, params("tag"))
   }
 }
