+++ src/main/scala/com/microsoft/partnercatalyst/fortis/spark/sources/streamprovider/ConnectorConfig.scala
@@ -1,3 +1,3 @@
-package com.microsoft.partnercatalyst.fortis.spark.streamprovider
+package com.microsoft.partnercatalyst.fortis.spark.sources.streamprovider
 
 case class ConnectorConfig(name: String, parameters: Map[String, String])
